# server: optimize chain interceptors (-1 allocation, -10% time/call)